requirements.txt
discord.py[voice]==2.3.2
yt-dlp==2023.11.16
spotipy==2.23.0
PyNaCl==1.5.0
aiohttp==3.9.1
asyncio==3.4.3